Refactor Dashboard and Notification Components
- Removed unused sorting functionality from Dashboard component for cleaner UI.
- Updated DashboardContentTable to eliminate the display of the modified date.
- Enhanced AuthButton to include a dropdown menu for user options and notifications.
- Replaced NotificationBell with a more streamlined NotificationsPanel for better notification handling.
